Updated Cashier, Branch , Item and Category

+++ pos-system/src/main/java/com/lifepill/possystem/dto/requestDTO/CashierUpdate/CashierRecentPinUpdateDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Data
 public class CashierRecentPinUpdateDTO {
-    private int cashierId;
+    private long cashierId;
     private int pin;
 }
