+++ pos-system/src/main/java/com/lifepill/possystem/dto/requestDTO/CashierUpdate/CashierUpdateDTO.java
@@ -11,7 +11,8 @@ import java.util.Date;
 @AllArgsConstructor
 @Data
 public class CashierUpdateDTO {
-    private int cashierId;
+    private long cashierId;
+//    private int branchId;
     private String cashierNicName;
     private String cashierEmail;
     private String cashierNic;
