+++ pos-system/src/main/java/com/lifepill/possystem/repo/cashierRepo/CashierBankDetailsRepo.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @EnableJpaRepositories
-public interface CashierBankDetailsRepo extends JpaRepository<CashierBankDetails, Integer> {
+public interface CashierBankDetailsRepo extends JpaRepository<CashierBankDetails, Long> {
 
 }
